Introduce WakeConvolution class
